Atlas/fra-atlas-dss/ai/requirements.txt
Flask==2.0.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
tensorflow==2.5.0
opencv-python==4.5.2.54
pytesseract==0.3.8
nltk==3.6.2
sqlalchemy==1.4.22
psycopg2-binary==2.9.1
geopandas==0.9.0
rasterio==1.2.6
matplotlib==3.4.2
seaborn==0.11.2
huggingface-hub==0.0.8